test(sentry): レビュー対応でテスト配置とdocstringを修正 (PR #532 review)
【背景】PR #532 のCodeRabbitレビューで、責務分割方針とsentinelテストの
docstring正確性に関する2件の指摘を受けた。

【変更内容】_before_send のイベント整形を検証するテストを
test_sentry_scrub_primitives.py から責務が一致する
test_sentry_scrub_events.py へ移動し、未使用となったimportを整理。
併せてsentinelテストのdocstringにあった編集残骸（文の途中切れ）を修復。

品質ゲート: 1443 tests passed / 97.58% coverage / ruff 0 errors / mypy 0 errors

Refs #532 (PR)
